refactor user procedure
