+++ src/server/api/routers/user/user.procedure.ts
@@ -1,4 +1,4 @@
-import { protectedProcedure, createTRPCRouter } from "../trpc";
+import { protectedProcedure, createTRPCRouter } from "../../trpc";
 
 export const userRouter = createTRPCRouter({
   get: protectedProcedure.query(({ ctx }) => ctx.user),
